refactor(metadata): decompose acl_exacl.rs into per-concern submodules
Split the 1050-line acl_exacl.rs into a directory of focused submodules:
- error.rs (is_unsupported_error)
- perms.rs (rsync<->exacl permission bit mapping)
- read.rs (get_rsync_acl)
- reset.rs (reset_acl_from_mode, clear_default_acl)
- reconstruct.rs (reconstruct_acl, rsync_acl_to_entries)
- sync.rs (sync_acls)
- apply.rs (apply_acls_from_cache)
- default_perms.rs (default_perms_for_dir)
- tests.rs (all original tests)

mod.rs re-exports the four public functions, preserving the existing
API surface byte-for-byte.
