--- crates/metadata/src/acl_exacl.rs
@@ -1,1050 +1,0 @@
-#![cfg(all(
-    feature = "acl",
-    any(target_os = "linux", target_os = "macos", target_os = "freebsd")
-))]
-
-//! Cross-platform ACL synchronization using the `exacl` crate.
-//!
-//! This module provides ACL synchronization for Linux (POSIX ACLs), macOS
-//! (extended ACLs), and FreeBSD (POSIX and NFSv4 ACLs) using the `exacl`
-//! crate for a unified, safe abstraction.
-//!
-//! # Design
-//!
-//! The [`sync_acls`] function coordinates the full ACL replication workflow:
-//!
-//! - Read access and default ACLs from the source without following symbolic
-//!   links unless explicitly requested.
-//! - Apply the retrieved ACLs to the destination.
-//! - When the source has no extended ACL entries, reset the destination to
-//!   match its permission bits.
-//!
-//! # Platform Differences
-//!
-//! - **Linux**: Uses POSIX ACLs with access and default ACL types.
-//! - **macOS**: Uses extended ACLs (NFSv4-style); no default ACLs on directories.
-//! - **FreeBSD**: Supports both POSIX and NFSv4 ACLs depending on filesystem.
-//!
-//! The `exacl` crate handles these differences internally.
-//!
-//! # Upstream Reference
-//!
-//! The behavior mirrors upstream rsync's ACL handling in `acls.c` and
-//! `lib/sysacls.c`, where:
-//! - ACL read/write errors on unsupported filesystems are silently ignored.
-//! - Symbolic links do not receive ACL updates (Linux doesn't support link ACLs).
-//! - Default ACLs are handled for directories on platforms that support them.
-//!
-//! # Examples
-//!
-//! ```rust,ignore
-//! use ::metadata::sync_acls;
-//! use std::path::Path;
-//!
-//! # fn demo() -> Result<(), ::metadata::MetadataError> {
-//! let source = Path::new("src");
-//! let destination = Path::new("dst");
-//! sync_acls(source, destination, true)?;
-//! # Ok(())
-//! # }
-//! ```
-
-#[cfg(any(target_os = "linux", target_os = "freebsd"))]
-use std::fs;
-use std::io;
-use std::path::Path;
-
-#[cfg(any(target_os = "linux", target_os = "freebsd"))]
-use exacl::AclOption;
-use exacl::{AclEntry, AclEntryKind, Perm, getfacl, setfacl};
-#[cfg(any(target_os = "linux", target_os = "freebsd"))]
-use protocol::acl::trace_default_perms_for_dir;
-use protocol::acl::{AclCache, IdAccess, NAME_IS_USER, NO_ENTRY, RsyncAcl};
-
-use crate::MetadataError;
-
-/// Synchronizes ACLs from `source` to `destination`.
-///
-/// Copies the access ACL and, when present on directories, the default ACL.
-/// When the source lacks extended ACL entries, the destination's ACL is reset
-/// to match its permission bits.
-///
-/// Symbolic links do not support ACLs; when `follow_symlinks` is `false`,
-/// this function returns immediately without performing any work.
-///
-/// # Errors
-///
-/// Returns [`MetadataError`] when reading the source ACLs or applying them
-/// to the destination fails. Filesystems that report ACLs as unsupported
-/// are treated as lacking ACLs and do not trigger an error.
-///
-/// # Upstream Reference
-///
-/// - `acls.c`: High-level ACL synchronization logic
-/// - `lib/sysacls.c`: Platform-specific ACL wrappers
-#[allow(clippy::module_name_repetitions)]
-pub fn sync_acls(
-    source: &Path,
-    destination: &Path,
-    follow_symlinks: bool,
-) -> Result<(), MetadataError> {
-    if !follow_symlinks {
-        return Ok(());
-    }
-
-    // Pre-check existence - ENOENT from getfacl would be masked by
-    // is_unsupported_error() which treats NotFound as "no ACL support".
-    if !source.exists() {
-        return Err(MetadataError::new(
-            "read ACL",
-            source,
-            io::Error::new(io::ErrorKind::NotFound, "source does not exist"),
-        ));
-    }
-
-    let source_acl = match getfacl(source, None) {
-        Ok(acl) => acl,
-        Err(e) => {
-            // upstream: acls.c - unsupported fs treated as empty ACL
-            if is_unsupported_error(&e) {
-                Vec::new()
-            } else {
-                return Err(MetadataError::new(
-                    "read ACL",
-                    source,
-                    io::Error::other(e.to_string()),
-                ));
-            }
-        }
-    };
-
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    let is_dir = match fs::symlink_metadata(source) {
-        Ok(m) => m.is_dir(),
-        Err(e) => return Err(MetadataError::new("stat", source, e)),
-    };
-
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    let default_acl = if is_dir {
-        match getfacl(source, Some(AclOption::DEFAULT_ACL)) {
-            Ok(acl) => Some(acl),
-            Err(e) if is_unsupported_error(&e) => None,
-            Err(e) => {
-                return Err(MetadataError::new(
-                    "read default ACL",
-                    source,
-                    io::Error::other(e.to_string()),
-                ));
-            }
-        }
-    } else {
-        None
-    };
-
-    if !source_acl.is_empty() {
-        if let Err(e) = setfacl(&[destination], &source_acl, None) {
-            if !is_unsupported_error(&e) {
-                return Err(MetadataError::new(
-                    "apply ACL",
-                    destination,
-                    io::Error::other(e.to_string()),
-                ));
-            }
-        }
-    } else {
-        reset_acl_from_mode(destination)?;
-    }
-
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    if is_dir {
-        match default_acl {
-            Some(acl) if !acl.is_empty() => {
-                if let Err(e) = setfacl(&[destination], &acl, Some(AclOption::DEFAULT_ACL)) {
-                    if !is_unsupported_error(&e) {
-                        return Err(MetadataError::new(
-                            "apply default ACL",
-                            destination,
-                            io::Error::other(e.to_string()),
-                        ));
-                    }
-                }
-            }
-            _ => {
-                clear_default_acl(destination)?;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-/// Converts [`exacl::Perm`] flags to rsync 3-bit rwx permission bits.
-fn exacl_perms_to_rsync(perms: Perm) -> u8 {
-    let mut bits: u8 = 0;
-    if perms.contains(Perm::READ) {
-        bits |= 0x04;
-    }
-    if perms.contains(Perm::WRITE) {
-        bits |= 0x02;
-    }
-    if perms.contains(Perm::EXECUTE) {
-        bits |= 0x01;
-    }
-    bits
-}
-
-/// Reads the filesystem ACL for `path` and converts it to an [`RsyncAcl`].
-///
-/// Returns a populated `RsyncAcl` with base entries (user_obj, group_obj,
-/// mask_obj, other_obj) and any named user/group entries. When the filesystem
-/// does not support ACLs or the path does not exist, returns a fake ACL
-/// derived from `mode` (matching upstream's fallback behavior).
-///
-/// # Arguments
-///
-/// * `path` - File or directory to read ACLs from
-/// * `mode` - File mode bits used as fallback when no ACL is available
-/// * `is_default` - If true, reads the default ACL (for directories only)
-///
-/// # Upstream Reference
-///
-/// Mirrors `get_rsync_acl()` in `acls.c` lines 472-536.
-pub fn get_rsync_acl(path: &Path, mode: u32, is_default: bool) -> RsyncAcl {
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    let option = if is_default {
-        Some(AclOption::DEFAULT_ACL)
-    } else {
-        Some(AclOption::ACCESS_ACL)
-    };
-    #[cfg(target_os = "macos")]
-    let option = {
-        let _ = is_default;
-        None
-    };
-
-    let entries = match getfacl(path, option) {
-        Ok(e) => e,
-        Err(e) => {
-            if is_unsupported_error(&e) {
-                // upstream: acls.c:525-528 - fake ACL from mode when unsupported
-                return if !is_default {
-                    RsyncAcl::from_mode(mode)
-                } else {
-                    RsyncAcl::new()
-                };
-            }
-            // upstream: acls.c:525-528 - mode-based fallback on error
-            return if !is_default {
-                RsyncAcl::from_mode(mode)
-            } else {
-                RsyncAcl::new()
-            };
-        }
-    };
-
-    if entries.is_empty() {
-        return if !is_default {
-            RsyncAcl::from_mode(mode)
-        } else {
-            RsyncAcl::new()
-        };
-    }
-
-    let mut acl = RsyncAcl::new();
-
-    for entry in &entries {
-        let perms = exacl_perms_to_rsync(entry.perms);
-        match entry.kind {
-            AclEntryKind::User if entry.name.is_empty() => {
-                // upstream: acls.c:481 - user_obj from unnamed User entry
-                if acl.user_obj == NO_ENTRY {
-                    acl.user_obj = perms;
-                }
-            }
-            AclEntryKind::Group if entry.name.is_empty() => {
-                // upstream: acls.c:484 - group_obj from unnamed Group entry
-                if acl.group_obj == NO_ENTRY {
-                    acl.group_obj = perms;
-                }
-            }
-            // upstream: acls.c:487-490 - mask and other POSIX entries
-            #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-            AclEntryKind::Mask => {
-                if acl.mask_obj == NO_ENTRY {
-                    acl.mask_obj = perms;
-                }
-            }
-            #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-            AclEntryKind::Other => {
-                if acl.other_obj == NO_ENTRY {
-                    acl.other_obj = perms;
-                }
-            }
-            AclEntryKind::User => {
-                // upstream: acls.c:497-501 - named user ACE
-                let uid: u32 = entry.name.parse().unwrap_or(0);
-                acl.names.push(IdAccess::user(uid, u32::from(perms)));
-            }
-            AclEntryKind::Group => {
-                // upstream: acls.c:502-506 - named group ACE
-                let gid: u32 = entry.name.parse().unwrap_or(0);
-                acl.names.push(IdAccess::group(gid, u32::from(perms)));
-            }
-            _ => {}
-        }
-    }
-
-    // upstream: acls.c:493-501 - fill NO_ENTRY base entries from mode
-    if acl.user_obj == NO_ENTRY {
-        acl.user_obj = ((mode >> 6) & 7) as u8;
-    }
-    if acl.group_obj == NO_ENTRY {
-        acl.group_obj = ((mode >> 3) & 7) as u8;
-    }
-    if acl.other_obj == NO_ENTRY {
-        acl.other_obj = (mode & 7) as u8;
-    }
-
-    acl
-}
-
-/// Resets the access ACL to match the file's permission bits.
-///
-/// On Linux and FreeBSD, converts the file's Unix permission mode into a
-/// minimal ACL using [`exacl::from_mode`]. On macOS, clears extended ACL
-/// entries by setting an empty ACL list.
-///
-/// # Errors
-///
-/// Returns [`MetadataError`] if reading file metadata or setting the ACL
-/// fails (unsupported filesystem errors are silently ignored).
-///
-/// # Upstream Reference
-///
-/// Matches upstream rsync's behavior when the source has no extended
-/// ACL entries - the destination retains only base owner/group/other entries.
-fn reset_acl_from_mode(path: &Path) -> Result<(), MetadataError> {
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    {
-        let metadata = match fs::symlink_metadata(path) {
-            Ok(m) => m,
-            Err(e) => return Err(MetadataError::new("stat", path, e)),
-        };
-
-        let mode = {
-            use std::os::unix::fs::PermissionsExt;
-            metadata.permissions().mode() & 0o777
-        };
-
-        let base_acl = exacl::from_mode(mode);
-
-        if let Err(e) = setfacl(&[path], &base_acl, None) {
-            if !is_unsupported_error(&e) {
-                return Err(MetadataError::new(
-                    "reset ACL",
-                    path,
-                    io::Error::other(e.to_string()),
-                ));
-            }
-        }
-    }
-
-    // macOS lacks exacl::from_mode - clear extended entries with empty list.
-    // Permission bits are managed separately from the extended ACL.
-    #[cfg(target_os = "macos")]
-    {
-        if let Err(e) = setfacl(&[path], &[], None) {
-            if !is_unsupported_error(&e) {
-                return Err(MetadataError::new(
-                    "reset ACL",
-                    path,
-                    io::Error::other(e.to_string()),
-                ));
-            }
-        }
-    }
-
-    Ok(())
-}
-
-/// Clears the default ACL from a directory (Linux/FreeBSD only).
-///
-/// Setting an empty default ACL removes it (equivalent to `setfacl -k`).
-/// Unsupported-filesystem errors are silently ignored.
-#[cfg(any(target_os = "linux", target_os = "freebsd"))]
-fn clear_default_acl(path: &Path) -> Result<(), MetadataError> {
-    if let Err(e) = setfacl(&[path], &[], Some(AclOption::DEFAULT_ACL)) {
-        if !is_unsupported_error(&e) {
-            return Err(MetadataError::new(
-                "clear default ACL",
-                path,
-                io::Error::other(e.to_string()),
-            ));
-        }
-    }
-    Ok(())
-}
-
-/// Returns `true` when an I/O error indicates an unsupported filesystem
-/// (FAT/VFAT, network mounts without ACL support, missing kernel xattrs).
-///
-/// Checks `ErrorKind`, then raw OS error codes (`ENOTSUP`, `ENOENT`, `EINVAL`,
-/// `ENODATA`, `EPERM`), then common error message substrings as a last resort.
-///
-/// upstream: `acls.c:no_acl_syscall_error()` - errors from filesystems that
-/// do not support ACLs are silently ignored.
-fn is_unsupported_error(e: &io::Error) -> bool {
-    matches!(
-        e.kind(),
-        io::ErrorKind::Unsupported | io::ErrorKind::InvalidInput | io::ErrorKind::NotFound
-    ) || {
-        match e.raw_os_error() {
-            Some(libc::ENOTSUP) | Some(libc::ENOENT) | Some(libc::EINVAL) | Some(libc::ENODATA)
-            | Some(libc::EPERM) => true,
-            _ => {
-                let msg = e.to_string().to_lowercase();
-                msg.contains("not supported")
-                    || msg.contains("no such file")
-                    || msg.contains("invalid argument")
-                    || msg.contains("no data")
-            }
-        }
-    }
-}
-
-/// Converts rsync permission bits (3-bit rwx) to [`exacl::Perm`] flags.
-fn rsync_perms_to_exacl(bits: u8) -> Perm {
-    let mut perms = Perm::empty();
-    if bits & 0x04 != 0 {
-        perms |= Perm::READ;
-    }
-    if bits & 0x02 != 0 {
-        perms |= Perm::WRITE;
-    }
-    if bits & 0x01 != 0 {
-        perms |= Perm::EXECUTE;
-    }
-    perms
-}
-
-/// Reconstructs a full ACL from stripped wire data and the file mode.
-///
-/// The sender strips base permission entries (user_obj, group_obj, other_obj)
-/// that can be inferred from the file mode. The receiver must reconstruct
-/// them before applying the ACL to the destination filesystem.
-///
-/// # Upstream Reference
-///
-/// Mirrors `change_sacl_perms()` in `acls.c` lines 857-933 which fills
-/// base entries from the file mode when the ACL was received stripped.
-fn reconstruct_acl(acl: &RsyncAcl, mode: Option<u32>) -> RsyncAcl {
-    let mode = match mode {
-        Some(m) => m,
-        None => return acl.clone(),
-    };
-
-    let mut result = acl.clone();
-
-    // upstream: acls.c:892 - user_obj from mode bits 8-6
-    if result.user_obj == NO_ENTRY {
-        result.user_obj = ((mode >> 6) & 7) as u8;
-    }
-    // upstream: acls.c:898 - group_obj from mode bits 5-3
-    if result.group_obj == NO_ENTRY {
-        result.group_obj = ((mode >> 3) & 7) as u8;
-    }
-    // upstream: acls.c:911 - other_obj from mode bits 2-0
-    if result.other_obj == NO_ENTRY {
-        result.other_obj = (mode & 7) as u8;
-    }
-    // upstream: acls.c:900-908 - mask from mode bits 5-3 when needed
-    if !result.names.is_empty() && result.mask_obj == NO_ENTRY {
-        result.mask_obj = ((mode >> 3) & 7) as u8;
-    }
-
-    result
-}
-
-/// Converts a [`RsyncAcl`] from the wire protocol into a list of [`AclEntry`]
-/// values suitable for [`exacl::setfacl`].
-///
-/// On Linux/FreeBSD, the resulting list contains POSIX ACL entries (user_obj,
-/// group_obj, mask, other, plus named user/group entries). On macOS, only
-/// named user/group entries are emitted as extended ACL entries since the
-/// base permissions are managed separately through file mode bits.
-///
-/// # Upstream Reference
-///
-/// Mirrors `set_rsync_acl()` in `acls.c` lines 835-928 which reconstructs
-/// a system ACL from the wire protocol `rsync_acl` struct.
-fn rsync_acl_to_entries(acl: &RsyncAcl) -> Vec<AclEntry> {
-    let mut entries = Vec::new();
-
-    // upstream: acls.c:866-878 - base entries for POSIX ACLs only (Linux/FreeBSD).
-    // macOS manages base mode bits separately from extended ACL entries.
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    {
-        if acl.has_user_obj() {
-            entries.push(AclEntry::allow_user(
-                "",
-                rsync_perms_to_exacl(acl.user_obj),
-                None,
-            ));
-        }
-        if acl.has_group_obj() {
-            entries.push(AclEntry::allow_group(
-                "",
-                rsync_perms_to_exacl(acl.group_obj),
-                None,
-            ));
-        }
-        if acl.has_mask_obj() {
-            entries.push(AclEntry::allow_mask(
-                rsync_perms_to_exacl(acl.mask_obj),
-                None,
-            ));
-        }
-        if acl.has_other_obj() {
-            entries.push(AclEntry::allow_other(
-                rsync_perms_to_exacl(acl.other_obj),
-                None,
-            ));
-        }
-    }
-
-    for ida in acl.names.iter() {
-        let perms = rsync_perms_to_exacl(ida.permissions() as u8);
-        let name = ida.id.to_string();
-
-        if ida.access & NAME_IS_USER != 0 {
-            entries.push(AclEntry::allow_user(&name, perms, None));
-        } else {
-            entries.push(AclEntry::allow_group(&name, perms, None));
-        }
-    }
-
-    entries
-}
-
-/// Applies parsed ACLs from an [`AclCache`] to a destination file.
-///
-/// This is the receiver-side function for applying ACLs that arrived over
-/// the wire protocol. The sender encodes ACLs during file list transmission
-/// and the receiver stores them in an [`AclCache`]. This function looks up
-/// the ACL by index and applies it to the destination path using `setfacl`.
-///
-/// For directories, both the access ACL and optional default ACL are applied.
-/// Symbolic links are skipped since they do not support ACLs on any platform.
-///
-/// # Arguments
-///
-/// * `destination` - Path to apply ACLs to.
-/// * `cache` - The ACL cache populated during file list reception.
-/// * `access_ndx` - Index into the access ACL cache.
-/// * `default_ndx` - Optional index into the default ACL cache (directories only).
-/// * `follow_symlinks` - Whether to follow symlinks. If `false`, returns immediately.
-///
-/// # Errors
-///
-/// Returns [`MetadataError`] if applying the ACL fails. Errors from filesystems
-/// that do not support ACLs are silently ignored.
-///
-/// # Upstream Reference
-///
-/// Mirrors `set_acl()` in `acls.c` lines 930-1001 which applies cached
-/// ACLs to destination files during the receiver's metadata application phase.
-#[allow(clippy::module_name_repetitions)]
-pub fn apply_acls_from_cache(
-    destination: &Path,
-    cache: &AclCache,
-    access_ndx: u32,
-    default_ndx: Option<u32>,
-    follow_symlinks: bool,
-    mode: Option<u32>,
-) -> Result<(), MetadataError> {
-    if !follow_symlinks {
-        return Ok(());
-    }
-
-    if let Some(acl) = cache.get_access(access_ndx) {
-        // upstream: acls.c:change_sacl_perms() - reconstruct base entries from mode
-        let reconstructed = reconstruct_acl(acl, mode);
-        let entries = rsync_acl_to_entries(&reconstructed);
-        if !entries.is_empty() {
-            if let Err(e) = setfacl(&[destination], &entries, None) {
-                if !is_unsupported_error(&e) {
-                    return Err(MetadataError::new(
-                        "apply ACL from cache",
-                        destination,
-                        io::Error::other(e.to_string()),
-                    ));
-                }
-            }
-        } else {
-            reset_acl_from_mode(destination)?;
-        }
-    }
-
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    if let Some(def_ndx) = default_ndx {
-        if let Some(def_acl) = cache.get_default(def_ndx) {
-            let entries = rsync_acl_to_entries(def_acl);
-            if !entries.is_empty() {
-                if let Err(e) = setfacl(&[destination], &entries, Some(AclOption::DEFAULT_ACL)) {
-                    if !is_unsupported_error(&e) {
-                        return Err(MetadataError::new(
-                            "apply default ACL from cache",
-                            destination,
-                            io::Error::other(e.to_string()),
-                        ));
-                    }
-                }
-            } else {
-                clear_default_acl(destination)?;
-            }
-        }
-    }
-
-    #[cfg(not(any(target_os = "linux", target_os = "freebsd")))]
-    let _ = default_ndx;
-
-    Ok(())
-}
-
-/// Computes the default permission bits a child of `dir` should inherit.
-///
-/// Mirrors upstream rsync's `default_perms_for_dir` (`acls.c:1083-1139`):
-///
-/// 1. Start from `ACCESSPERMS & ~umask` (the POSIX default).
-/// 2. Read the directory's default POSIX ACL.
-/// 3. When the ACL has a `user_obj` entry, fold its bits into the returned
-///    permission mask via `rsync_acl_get_perms`.
-/// 4. When `--debug=ACL` is at level 1, emit `got ACL-based default perms %o
-///    for directory %s` (`acls.c:1133-1134`).
-///
-/// Unsupported filesystems and missing default ACLs both fall back to the
-/// umask-derived default without emitting; only the successful ACL lookup
-/// produces the upstream debug line.
-///
-/// On platforms that do not expose POSIX default ACLs (e.g. macOS), this
-/// function returns the umask-derived default and never emits, matching the
-/// `#ifdef SUPPORT_ACLS` gating at `generator.c:1337-1340`.
-///
-/// # Upstream Reference
-///
-/// - `acls.c:1083-1139` `default_perms_for_dir`
-/// - `generator.c:1337-1340` and `receiver.c:846-851` - the two call sites
-///   that fold the returned bits into `dest_mode()` when `--perms` is off.
-#[allow(clippy::module_name_repetitions)]
-#[must_use]
-pub fn default_perms_for_dir(dir: &Path, orig_umask: u32) -> u32 {
-    // upstream: acls.c:1093 - perms = ACCESSPERMS & ~orig_umask
-    let default_perms = 0o777u32 & !(orig_umask & 0o777);
-
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    {
-        let dir_label = dir.to_string_lossy();
-        let entries = match getfacl(dir, Some(AclOption::DEFAULT_ACL)) {
-            Ok(e) => e,
-            Err(_) => return default_perms,
-        };
-
-        if entries.is_empty() {
-            return default_perms;
-        }
-
-        // upstream: acls.c:1131 - racl.user_obj != NO_ENTRY guard
-        let mut user_obj_perms: Option<u8> = None;
-        let mut group_obj_perms: Option<u8> = None;
-        let mut other_obj_perms: Option<u8> = None;
-        for entry in &entries {
-            let perms = exacl_perms_to_rsync(entry.perms);
-            match entry.kind {
-                AclEntryKind::User if entry.name.is_empty() && user_obj_perms.is_none() => {
-                    user_obj_perms = Some(perms);
-                }
-                AclEntryKind::Group if entry.name.is_empty() && group_obj_perms.is_none() => {
-                    group_obj_perms = Some(perms);
-                }
-                AclEntryKind::Other if other_obj_perms.is_none() => {
-                    other_obj_perms = Some(perms);
-                }
-                _ => {}
-            }
-        }
-
-        let Some(user_obj) = user_obj_perms else {
-            return default_perms;
-        };
-        // upstream: acls.c:1132 - perms = rsync_acl_get_perms(&racl)
-        //   = (user_obj << 6) | (group_obj << 3) | other_obj
-        let perms = (u32::from(user_obj) << 6)
-            | (u32::from(group_obj_perms.unwrap_or(0)) << 3)
-            | u32::from(other_obj_perms.unwrap_or(0));
-        // upstream: acls.c:1133-1134 - DEBUG_GTE(ACL, 1) emission
-        trace_default_perms_for_dir(perms, &dir_label);
-        perms
-    }
-
-    #[cfg(not(any(target_os = "linux", target_os = "freebsd")))]
-    {
-        // macOS and other targets without POSIX default ACLs: no emission,
-        // matches upstream's `#ifdef SUPPORT_ACLS` guard at generator.c:1337.
-        let _ = dir;
-        default_perms
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use exacl::AclEntryKind;
-    use std::fs::File;
-    use tempfile::tempdir;
-
-    #[test]
-    fn sync_acls_skips_when_not_following_symlinks() {
-        let dir = tempdir().expect("tempdir");
-        let source = dir.path().join("src");
-        let destination = dir.path().join("dst");
-        File::create(&source).expect("create src");
-        File::create(&destination).expect("create dst");
-
-        // Should return Ok without doing anything
-        let result = sync_acls(&source, &destination, false);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn sync_acls_copies_between_regular_files() {
-        let dir = tempdir().expect("tempdir");
-        let source = dir.path().join("src");
-        let destination = dir.path().join("dst");
-        File::create(&source).expect("create src");
-        File::create(&destination).expect("create dst");
-
-        // Should succeed for files on same filesystem
-        let result = sync_acls(&source, &destination, true);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn sync_acls_works_with_directories() {
-        let dir = tempdir().expect("tempdir");
-        let source = dir.path().join("src_dir");
-        let destination = dir.path().join("dst_dir");
-        std::fs::create_dir(&source).expect("create src_dir");
-        std::fs::create_dir(&destination).expect("create dst_dir");
-
-        let result = sync_acls(&source, &destination, true);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn reset_acl_from_mode_works() {
-        let dir = tempdir().expect("tempdir");
-        let file = dir.path().join("test");
-        File::create(&file).expect("create file");
-
-        let result = reset_acl_from_mode(&file);
-        assert!(result.is_ok());
-    }
-
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    #[test]
-    fn clear_default_acl_works_on_directory() {
-        let dir = tempdir().expect("tempdir");
-        let subdir = dir.path().join("subdir");
-        std::fs::create_dir(&subdir).expect("create subdir");
-
-        let result = clear_default_acl(&subdir);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn is_unsupported_error_detects_common_messages() {
-        // Test various error message patterns
-        let patterns = [
-            "operation not supported",
-            "No such file or directory",
-            "Invalid argument",
-            "No data available",
-        ];
-
-        for pattern in patterns {
-            let err = std::io::Error::other(pattern);
-            assert!(
-                is_unsupported_error(&err),
-                "should detect '{pattern}' as unsupported"
-            );
-        }
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn is_unsupported_error_detects_os_error_codes() {
-        // Test raw OS error codes
-        let codes = [libc::ENOTSUP, libc::ENOENT, libc::EINVAL, libc::EPERM];
-
-        for code in codes {
-            let err = std::io::Error::from_raw_os_error(code);
-            assert!(
-                is_unsupported_error(&err),
-                "should detect OS error code {code} as unsupported"
-            );
-        }
-    }
-
-    #[test]
-    fn rsync_perms_to_exacl_all_bits() {
-        assert_eq!(rsync_perms_to_exacl(0x00), Perm::empty());
-        assert_eq!(rsync_perms_to_exacl(0x01), Perm::EXECUTE);
-        assert_eq!(rsync_perms_to_exacl(0x02), Perm::WRITE);
-        assert_eq!(rsync_perms_to_exacl(0x04), Perm::READ);
-        assert_eq!(
-            rsync_perms_to_exacl(0x07),
-            Perm::READ | Perm::WRITE | Perm::EXECUTE
-        );
-        assert_eq!(rsync_perms_to_exacl(0x05), Perm::READ | Perm::EXECUTE);
-    }
-
-    #[test]
-    fn rsync_acl_to_entries_empty_acl() {
-        let acl = RsyncAcl::new();
-        let entries = rsync_acl_to_entries(&acl);
-        assert!(entries.is_empty());
-    }
-
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    #[test]
-    fn rsync_acl_to_entries_base_entries() {
-        let acl = RsyncAcl::from_mode(0o754);
-        let entries = rsync_acl_to_entries(&acl);
-
-        // user_obj(rwx) + group_obj(r-x) + other_obj(r--)
-        assert_eq!(entries.len(), 3);
-        assert_eq!(entries[0].kind, AclEntryKind::User);
-        assert_eq!(entries[0].name, "");
-        assert_eq!(entries[0].perms, Perm::READ | Perm::WRITE | Perm::EXECUTE);
-        assert_eq!(entries[1].kind, AclEntryKind::Group);
-        assert_eq!(entries[1].name, "");
-        assert_eq!(entries[1].perms, Perm::READ | Perm::EXECUTE);
-        assert_eq!(entries[2].kind, AclEntryKind::Other);
-        assert_eq!(entries[2].name, "");
-        assert_eq!(entries[2].perms, Perm::READ);
-    }
-
-    #[test]
-    fn rsync_acl_to_entries_named_user_and_group() {
-        use protocol::acl::IdAccess;
-
-        let mut acl = RsyncAcl::from_mode(0o755);
-        acl.names.push(IdAccess::user(1000, 0x07));
-        acl.names.push(IdAccess::group(100, 0x05));
-
-        let entries = rsync_acl_to_entries(&acl);
-
-        // Find named entries (skip base entries on Linux/FreeBSD)
-        let named: Vec<_> = entries.iter().filter(|e| !e.name.is_empty()).collect();
-        assert_eq!(named.len(), 2);
-        assert_eq!(named[0].kind, AclEntryKind::User);
-        assert_eq!(named[0].name, "1000");
-        assert_eq!(named[0].perms, Perm::READ | Perm::WRITE | Perm::EXECUTE);
-        assert_eq!(named[1].kind, AclEntryKind::Group);
-        assert_eq!(named[1].name, "100");
-        assert_eq!(named[1].perms, Perm::READ | Perm::EXECUTE);
-    }
-
-    #[test]
-    fn apply_acls_from_cache_skips_symlinks() {
-        let dir = tempdir().expect("tempdir");
-        let file = dir.path().join("test");
-        File::create(&file).expect("create file");
-
-        let cache = AclCache::new();
-        let result = apply_acls_from_cache(&file, &cache, 0, None, false, None);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn apply_acls_from_cache_applies_basic_acl() {
-        let dir = tempdir().expect("tempdir");
-        let file = dir.path().join("test");
-        File::create(&file).expect("create file");
-
-        let mut cache = AclCache::new();
-        let acl = RsyncAcl::from_mode(0o644);
-        let ndx = cache.store_access(acl);
-
-        let result = apply_acls_from_cache(&file, &cache, ndx, None, true, Some(0o644));
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn apply_acls_from_cache_empty_acl_resets() {
-        let dir = tempdir().expect("tempdir");
-        let file = dir.path().join("test");
-        File::create(&file).expect("create file");
-
-        let mut cache = AclCache::new();
-        let acl = RsyncAcl::new();
-        let ndx = cache.store_access(acl);
-
-        let result = apply_acls_from_cache(&file, &cache, ndx, None, true, Some(0o644));
-        assert!(result.is_ok());
-    }
-
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    #[test]
-    fn apply_acls_from_cache_directory_with_default() {
-        let dir = tempdir().expect("tempdir");
-        let subdir = dir.path().join("subdir");
-        std::fs::create_dir(&subdir).expect("create subdir");
-
-        let mut cache = AclCache::new();
-        let access = RsyncAcl::from_mode(0o755);
-        let default = RsyncAcl::from_mode(0o755);
-        let access_ndx = cache.store_access(access);
-        let default_ndx = cache.store_default(default);
-
-        let result = apply_acls_from_cache(
-            &subdir,
-            &cache,
-            access_ndx,
-            Some(default_ndx),
-            true,
-            Some(0o755),
-        );
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn apply_acls_from_cache_missing_index_is_noop() {
-        let dir = tempdir().expect("tempdir");
-        let file = dir.path().join("test");
-        File::create(&file).expect("create file");
-
-        let cache = AclCache::new();
-        // Index 99 does not exist - should be a no-op, not an error
-        let result = apply_acls_from_cache(&file, &cache, 99, None, true, Some(0o644));
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn default_perms_for_dir_no_acl_returns_umask_default() {
-        // No default ACL: upstream returns ACCESSPERMS & ~orig_umask without
-        // emitting `DEBUG_GTE(ACL, 1)`.
-        let dir = tempdir().expect("tempdir");
-        let target = dir.path().join("no_default_acl");
-        std::fs::create_dir(&target).expect("create dir");
-
-        assert_eq!(default_perms_for_dir(&target, 0o022), 0o755);
-        assert_eq!(default_perms_for_dir(&target, 0o077), 0o700);
-        assert_eq!(default_perms_for_dir(&target, 0), 0o777);
-    }
-
-    #[test]
-    fn default_perms_for_dir_missing_dir_returns_umask_default() {
-        // getfacl error path: upstream falls back to umask-derived default
-        // and never emits. Verify the same here.
-        let dir = tempdir().expect("tempdir");
-        let missing = dir.path().join("nonexistent");
-
-        assert_eq!(default_perms_for_dir(&missing, 0o022), 0o755);
-    }
-
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    #[test]
-    fn default_perms_for_dir_emits_when_default_acl_present() {
-        // upstream: acls.c:1131-1134 - DEBUG_GTE(ACL, 1) fires when the
-        // directory's default ACL unpacked into a user_obj entry.
-        use logging::{DebugFlag, DiagnosticEvent, VerbosityConfig, drain_events, init};
-
-        let dir = tempdir().expect("tempdir");
-        let target = dir.path().join("with_default_acl");
-        std::fs::create_dir(&target).expect("create dir");
-
-        // Install a default ACL: user::rwx, group::r-x, other::r-x.
-        let default_entries = vec![
-            exacl::AclEntry::allow_user("", Perm::READ | Perm::WRITE | Perm::EXECUTE, None),
-            exacl::AclEntry::allow_group("", Perm::READ | Perm::EXECUTE, None),
-            exacl::AclEntry::allow_other(Perm::READ | Perm::EXECUTE, None),
-        ];
-        if setfacl(&[&target], &default_entries, Some(AclOption::DEFAULT_ACL)).is_err() {
-            // Filesystem doesn't support default ACLs (e.g., tmpfs without acl mount).
-            // Skip the emission assertion; this is the same fallback upstream takes.
-            return;
-        }
-
-        let mut cfg = VerbosityConfig::default();
-        cfg.debug.acl = 1;
-        init(cfg);
-        let _ = drain_events();
-
-        let perms = default_perms_for_dir(&target, 0o022);
-        assert_eq!(perms, 0o755);
-
-        let messages: Vec<String> = drain_events()
-            .into_iter()
-            .filter_map(|event| match event {
-                DiagnosticEvent::Debug {
-                    flag: DebugFlag::Acl,
-                    message,
-                    ..
-                } => Some(message),
-                _ => None,
-            })
-            .collect();
-        let expected = format!(
-            "got ACL-based default perms 755 for directory {}",
-            target.display()
-        );
-        assert!(
-            messages.iter().any(|m| m == &expected),
-            "expected emission {expected:?}, got {messages:?}"
-        );
-    }
-
-    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
-    #[test]
-    fn default_perms_for_dir_no_emission_when_disabled() {
-        // upstream: DEBUG_GTE(ACL, 1) is gated; level 0 suppresses the line
-        // even when the default ACL unpacks successfully.
-        use logging::{DebugFlag, DiagnosticEvent, VerbosityConfig, drain_events, init};
-
-        let dir = tempdir().expect("tempdir");
-        let target = dir.path().join("with_default_acl_silent");
-        std::fs::create_dir(&target).expect("create dir");
-
-        let default_entries = vec![
-            exacl::AclEntry::allow_user("", Perm::READ | Perm::WRITE | Perm::EXECUTE, None),
-            exacl::AclEntry::allow_group("", Perm::READ, None),
-            exacl::AclEntry::allow_other(Perm::empty(), None),
-        ];
-        if setfacl(&[&target], &default_entries, Some(AclOption::DEFAULT_ACL)).is_err() {
-            return;
-        }
-
-        let mut cfg = VerbosityConfig::default();
-        cfg.debug.acl = 0;
-        init(cfg);
-        let _ = drain_events();
-
-        let _ = default_perms_for_dir(&target, 0o022);
-        let messages: Vec<String> = drain_events()
-            .into_iter()
-            .filter_map(|event| match event {
-                DiagnosticEvent::Debug {
-                    flag: DebugFlag::Acl,
-                    message,
-                    ..
-                } => Some(message),
-                _ => None,
-            })
-            .collect();
-        assert!(
-            messages.is_empty(),
-            "ACL debug must be suppressed at level 0, got {messages:?}"
-        );
-    }
-}
